Better participant add messaging, support remove
