+++ adapters/index.js
@@ -2,6 +2,7 @@ module.exports = {
 	'ReviewCreatedFeedEventBean': require('./review-created-feed-event'),
 	'ReviewStateChangedFeedEventBean': require('./review-state-changed-feed-event'),
 	'ParticipantStateChangedFeedEventBean': require('./participant-state-changed-feed-event'),
-	'NewParticipantInReviewFeedEventBean': require('./new-participant-in-review-feed-event'),
+	'NewParticipantInReviewFeedEventBean': require('./participant-in-review-feed-event'),
+	'RemovedParticipantFromReviewFeedEventBean': require('./participant-in-review-feed-event'),
 	'DiscussionFeedEventBean': require('./discussion-feed-event')
 };
